server-python/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
requests==2.31.0
stripe==8.5.0
google-generativeai==0.4.1
algoliasearch==3.0.0
